+++ developer_resources/test_actual_users.py
@@ -69,7 +69,8 @@ def run_with_timeout(cmd):
     num_attempts = 0
     while num_attempts < 5:
         try:
-            output = infra_utils.run_cmd(cmd, timeout=5)
+            result = infra_utils.run_cmd(cmd, timeout=5)
+            output = result.stdout
         except subprocess.TimeoutExpired:
             print("Time out.")
             num_attempts += 1