+++ src/gh_profiler/utils/infra_utils.py
@@ -3,6 +3,8 @@
 import os
 import shlex
 import subprocess
+from dataclasses import dataclass
+
 
 DEFAULT_ENV = {
     **os.environ,
@@ -10,9 +12,21 @@ DEFAULT_ENV = {
     "NO_COLOR": "1",
 }
 
+@dataclass
+class CommandResult:
+    """Decoded subprocess output."""
+    stdout: str
+    stderr: str
+    returncode: int
+
 
 def run_cmd(cmd, env=DEFAULT_ENV, timeout=None):
-    """Run a subprocess command, return stdout."""
+    """Run a subprocess command, return CommandResult instance."""
     cmd_parts = shlex.split(cmd)
     output_obj = subprocess.run(cmd_parts, capture_output=True, env=env, timeout=timeout)
-    return output_obj.stdout.decode()
+
+    return CommandResult(
+        stdout=output_obj.stdout.decode(),
+        stderr=output_obj.stderr.decode(),
+        returncode=output_obj.returncode,
+    )